chore: updated wrapper name, exported it from entities-saved-states & updated preview panel to use wrapper
