+++ packages/editor/src/components/save-publish-panels/index.js
@@ -5,16 +5,12 @@ import { useSelect, useDispatch } from '@wordpress/data';
 import { Button, createSlotFill } from '@wordpress/components';
 import { __ } from '@wordpress/i18n';
 import { useCallback } from '@wordpress/element';
-import {
-	__experimentalUseDialog as useDialog,
-	useInstanceId,
-} from '@wordpress/compose';
 
 /**
  * Internal dependencies
  */
 import EntitiesSavedStates, {
-	EntitiesSavedStatesExtensible,
+	EntitiesSavedStatesDialogWrapper,
 } from '../entities-saved-states';
 import PostPublishPanel from '../post-publish-panel';
 import PluginPrePublishPanel from '../plugin-pre-publish-panel';
@@ -108,47 +104,14 @@ export default function SavePublishPanels( {
 	return (
 		<>
 			{ isEntitiesSavedStatesOpen && (
-				<EntitieDialogWrapper close={ closeEntitiesSavedStates }>
+				<EntitiesSavedStatesDialogWrapper
+					close={ closeEntitiesSavedStates }
+				>
 					<EntitiesSavedStates close={ closeEntitiesSavedStates } />
-				</EntitieDialogWrapper>
+				</EntitiesSavedStatesDialogWrapper>
 			) }
 			<Slot bubblesVirtually />
 			{ ! isEntitiesSavedStatesOpen && unmountableContent }
 		</>
-	);
-}
-
-/**
- * A wrapper component that renders a dialog for displaying entities.
- *
- * @param {Object}          props          The component's props.
- * @param {React.ReactNode} props.children The content to be displayed inside the dialog.
- * @param {Function}        props.close    A function to close the dialog.
- *
- * @return {React.Element} The rendered dialog element with children.
- */
-export function EntitieDialogWrapper( { children, close } ) {
-	const dismissPanel = useCallback( () => close(), [ close ] );
-	const [ saveDialogRef, saveDialogProps ] = useDialog( {
-		onClose: () => dismissPanel(),
-	} );
-
-	const dialogLabel = useInstanceId( EntitiesSavedStatesExtensible, 'label' );
-	const dialogDescription = useInstanceId(
-		EntitiesSavedStatesExtensible,
-		'description'
-	);
-
-	return (
-		<div
-			ref={ saveDialogRef }
-			{ ...saveDialogProps }
-			role="dialog"
-			aria-labelledby={ dialogLabel }
-			aria-describedby={ dialogDescription }
-		>
-			{ ' ' }
-			{ children }
-		</div>
 	);
 }
